Add: move to trash/restore from trash functionality. LA

+++ src/main/java/com/lisaanna/ws_todo/service/DeletedTaskService.java
@@ -33,20 +33,24 @@ public class DeletedTaskService {
     }
 
     // DELETE - by id
-    public void deleteTaskFromTrash(String id) {
-
+    public boolean deleteTaskFromTrash(String id) {
         DeletedTask taskToDelete = deletedTaskRepository.findById(id).orElse(null);
 
         if (taskToDelete != null) {
             deletedTaskRepository.delete(taskToDelete);
+            return true;
         } else {
-            throw new NullPointerException();
+            return false;
         }
-
     }
 
     // DELETE - all
-    public void deleteAllTasks() {
-        deletedTaskRepository.deleteAll();
+    public boolean deleteAllTasks() {
+        if (deletedTaskRepository.findAll().isEmpty()) {
+            return false;
+        } else {
+            deletedTaskRepository.deleteAll();
+            return true;
+        }
     }
 }
